fix: when changing session's profile, named profile in credentials ini file will be modified only if the session is active
refers to #110, #109 and #107

+++ src/app/session/session-card/session-card.component.ts
@@ -319,7 +319,9 @@ export class SessionCardComponent extends AntiMemLeak implements OnInit {
 
   changeDefaultProfile() {
     if (this.selectedProfile) {
-      this.sessionService.removeFromIniFile(this.session.profile);
+      if (this.session.active) {
+        this.sessionService.removeFromIniFile(this.session.profile);
+      }
       this.sessionService.addProfile(this.selectedProfile);
       this.sessionService.updateSessionProfile(this.session, this.selectedProfile);
